# backend/docs/swagger.yaml
definitions:
...
  protocol.AppleWebApplyBundleIDReq:
    properties:
      bundleId:
        description: Bundle ID
        example: cn.ivfzhou.test
        maxLength: 64
        type: string
      type:
        description: 类型
        example: 1
        maximum: 2
        minimum: 1
        type: integer
    required:
    - bundleId
    type: object
  protocol.AppleWebApplyCertificateReq:
    properties:
      type:
        description: 类型
        example: IOS_DEVELOPMENT
        type: string
    type: object
  protocol.AppleWebApplyInHouseProfileReq:
    properties:
      bundleId:
        description: Bundle ID
        example: cn.ivfzhou.test
        maxLength: 64
        type: string
    required:
    - bundleId
    type: object
  protocol.AppleWebApplyProfileReq:
    properties:
      bundleId:
        description: Bundle ID
        example: cn.ivfzhou.test
        maxLength: 64
        type: string
      platform:
        description: 平台
        example: IOS
        type: string
      type:
        description: 类型
        example: IOS_APP_DEVELOPMENT
        type: string
    required:
    - bundleId
    type: object
...
  util.Response-any:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
host: 127.0.0.1:8090
info:
  contact:
    email: ivfzhou@126.com
    name: ivfzhou
  license:
    name: MulanPSL-2.0 license
    url: http://license.coscl.org.cn/MulanPSL2
  title: CSMS API
  version: "1.0"
paths:
...
  /web/apple/applyBundleID/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleWebApplyBundleIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请 Bundle ID
      tags:
      - Apple-WebAPI
  /web/apple/applyCertificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleWebApplyCertificateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请苹果签名证书
      tags:
      - Apple-WebAPI
  /web/apple/applyCommonProfile/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请企业内测通配符描述文件
      tags:
      - Apple-WebAPI
  /web/apple/applyInHouseProfile/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleWebApplyInHouseProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请企业内测描述文件
      tags:
      - Apple-WebAPI
  /web/apple/applyProfile/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleWebApplyProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请描述文件
      tags:
      - Apple-WebAPI
  /web/apple/applyPushCertificate/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleWebApplyPushCertificateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 申请 Push 证书
      tags:
      - Apple-WebAPI
  /web/apple/deleteBundleID/{appId}:
    delete:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: Bundle ID。
        example: cn.ivfzhou.test
        in: query
        maxLength: 64
        name: bundleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 删除 Bundle ID
      tags:
      - Apple-WebAPI
  /web/apple/downloadCertificate/{appId}:
    get:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        in: query
        name: certificateId
        type: string
      - description: 类型
        example: 1
        in: query
        maximum: 3
        minimum: 1
        name: type
        type: integer
      produces:
      - application/otect-stream
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              type: string
          schema:
            type: file
      summary: 下载证书和描述文件
      tags:
      - Apple-WebAPI
  /web/apple/listAppCertificates/{appId}:
    get:
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleWebListAppCertificatesRsp'
      summary: 获取应用证书和描述文件列表
      tags:
      - Apple-WebAPI
  /web/apple/listBundleIDs/{appId}:
    get:
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleWebListBundleIDsRsp'
      summary: 获取 Bundle ID 列表
      tags:
      - Apple-WebAPI
  /web/apple/listCertificates:
    get:
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleWebListCertificatesRsp'
      summary: 获取苹果证书信息列表
      tags:
      - Apple-WebAPI
  /web/apple/listDevices/{appId}:
    get:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 页数
        example: 1
        in: query
        minimum: 1
        name: pageNumber
        type: integer
      - description: 一页任务条数
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleWebListDevicesRsp'
      summary: 设备列表
      tags:
      - Apple-WebAPI
...
  /web/apple/removeCertificate/{appId}:
    delete:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 证书 ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        in: query
        name: certificateId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 删除苹果证书
      tags:
      - Apple-WebAPI
...
schemes:
- http
swagger: "2.0"